==> 2.Feature Engineering/1_members processing.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
import warnings

df_member = pd.read_csv('../0.Data/Members.csv')

df_member.dtypes

df_member.head()

#Count missing data
def count_missing(df):   
    total = df.isnull().sum().sort_values(ascending=False)
    percent = (df.isnull().sum()/df.isnull().count()).sort_values(ascending=False)
    missing_data  = pd.concat([total, percent], axis=1, keys=['Total', 'Percent'])
    print(missing_data.head(20))
    return missing_data

count_missing(df_member)

# One-Hot to Male, Female and MissSex
df_member['Male'] = 0
df_member['Female'] = 0
df_member['MissSex'] = 0
df_member.loc[df_member.Sex == 'M', 'Male'] = 1
df_member.loc[df_member.Sex == 'F', 'Female'] = 1
df_member.loc[df_member.Sex.isna(), 'MissSex'] = 1
df_member.drop(columns=['Sex'], inplace=True)

# Replace Age with mean
df_member['MissAge'] = 0
df_member.loc[df_member.AgeAtFirstClaim == '0-9', 'AgeAtFirstClaim'] = 5
df_member.loc[df_member.AgeAtFirstClaim == '10-19', 'AgeAtFirstClaim'] = 15
df_member.loc[df_member.AgeAtFirstClaim == '20-29', 'AgeAtFirstClaim'] = 25
df_member.loc[df_member.AgeAtFirstClaim == '30-39', 'AgeAtFirstClaim'] = 35
df_member.loc[df_member.AgeAtFirstClaim == '40-49', 'AgeAtFirstClaim'] = 45
df_member.loc[df_member.AgeAtFirstClaim == '50-59', 'AgeAtFirstClaim'] = 55
df_member.loc[df_member.AgeAtFirstClaim == '60-69', 'AgeAtFirstClaim'] = 65
df_member.loc[df_member.AgeAtFirstClaim == '70-79', 'AgeAtFirstClaim'] = 75
df_member.loc[df_member.AgeAtFirstClaim == '80+', 'AgeAtFirstClaim'] = 80
df_member.loc[df_member.AgeAtFirstClaim.isna(), 'AgeAtFirstClaim'] = 45 # Average
df_member.loc[df_member.AgeAtFirstClaim.isna(), 'MissAge'] = 1

# Save to CSV file
df_member.to_csv(path_or_buf='New Data/df_members_new.csv', index=False)

df_member.describe

df_member.info()
