==> 2.Feature Engineering/2_Claims processing.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
import warnings
from sklearn.preprocessing import MultiLabelBinarizer

df_claims = pd.read_csv('../0.Data/Claims.csv')

df_claims.dtypes

df_claims.head()

df_claims.loc[(df_claims['MemberID']==33120) & (df_claims['Year']=='Y2')].shape

df_claims.info()

#Count missing data
def count_missing(df):   
    total = df.isnull().sum().sort_values(ascending=False)
    percent = (df.isnull().sum()/df.isnull().count()).sort_values(ascending=False)
    missing_data  = pd.concat([total, percent], axis=1, keys=['Total', 'Percent'])
    print(missing_data.head(20))
    return missing_data

count_missing(df_claims)

#Update paydelay
df_claims.PayDelay.replace({'162+': '162'}, inplace=True)
df_claims['PayDelay'] = df_claims['PayDelay'].astype(int)

df_claims.dtypes

#Update DSFS
df_claims.loc[df_claims['DSFS']=='0- 1 month','DSFS'] = 1
df_claims.loc[df_claims['DSFS']=='1- 2 months','DSFS'] = 2
df_claims.loc[df_claims['DSFS']=='2- 3 months','DSFS'] = 3
df_claims.loc[df_claims['DSFS']=='3- 4 months','DSFS'] = 4
df_claims.loc[df_claims['DSFS']=='4- 5 months','DSFS'] = 5
df_claims.loc[df_claims['DSFS']=='5- 6 months','DSFS'] = 6
df_claims.loc[df_claims['DSFS']=='6- 7 months','DSFS'] = 7
df_claims.loc[df_claims['DSFS']=='7- 8 months','DSFS'] = 8
df_claims.loc[df_claims['DSFS']=='8- 9 months','DSFS'] = 9
df_claims.loc[df_claims['DSFS']=='9-10 months','DSFS'] = 10
df_claims.loc[df_claims['DSFS']=='10-11 months','DSFS'] = 11
df_claims.loc[df_claims['DSFS']=='11-12 months','DSFS'] = 12
df_claims['DSFS'].fillna(0, inplace=True)
df_claims['DSFS'] = df_claims['DSFS'].astype(int)

df_claims.dtypes

#Update Charlson
df_claims.loc[df_claims['CharlsonIndex']=='0','CharlsonIndex'] = 0
df_claims.loc[df_claims['CharlsonIndex']=='1-2','CharlsonIndex'] = 2
df_claims.loc[df_claims['CharlsonIndex']=='3-4','CharlsonIndex'] = 4
df_claims.loc[df_claims['CharlsonIndex']=='5+','CharlsonIndex'] = 6
df_claims['CharlsonIndex'] = df_claims['CharlsonIndex'].astype(int)

df_claims.dtypes

df_claims['MemberID'] = df_claims['MemberID'].astype(object)
df_claims['ProviderID'] = df_claims['ProviderID'].astype(object)
#convert Vendor từ float64 thành object
df_claims['Vendor'] = df_claims['Vendor'].astype(object)
#convert PCP từ float64 thành object
df_claims['PCP'] = df_claims['PCP'].astype(object)

df_claims.dtypes

def impute_nan_create_category(DataFrame,ColName):
     DataFrame[ColName] = np.where(DataFrame[ColName].isnull(),"Unknown",DataFrame[ColName])

#impute ProviderID
impute_nan_create_category(df_claims,'ProviderID')

#impute Vendor
impute_nan_create_category(df_claims,'Vendor')

#impute PCP
impute_nan_create_category(df_claims,'PCP')

#impute Specialty
impute_nan_create_category(df_claims,'Specialty')

#impute PlaceSvc
impute_nan_create_category(df_claims,'PlaceSvc')

#impute PrimaryConditionGroup 
impute_nan_create_category(df_claims,'PrimaryConditionGroup')

#impute ProcedureGroup 
impute_nan_create_category(df_claims,'ProcedureGroup')

count_missing(df_claims)

#Update LengthOfStay
df_claims['LOS_UNKNOWN'] = 0
df_claims['LOS_SUPRESSED'] = 0
df_claims['LOS_KNOWN'] = 0
m = (df_claims['LengthOfStay'].isna()) & (df_claims['SupLOS'] == 0)
df_claims.loc[m, 'LOS_UNKNOWN'] = 1
m = (df_claims['LengthOfStay'].isna()) & (df_claims['SupLOS'] == 1)
df_claims.loc[m, 'LOS_SUPRESSED'] = 1
m = (df_claims['LengthOfStay'].notna()) & (df_claims['SupLOS'] == 1)
df_claims.loc[m, 'LOS_KNOWN'] = 1

df_claims.loc[df_claims['LengthOfStay']=='1 day','LengthOfStay'] = 1
df_claims.loc[df_claims['LengthOfStay']=='2 days','LengthOfStay'] = 2
df_claims.loc[df_claims['LengthOfStay']=='3 days','LengthOfStay'] = 3
df_claims.loc[df_claims['LengthOfStay']=='4 days','LengthOfStay'] = 4
df_claims.loc[df_claims['LengthOfStay']=='5 days','LengthOfStay'] = 5
df_claims.loc[df_claims['LengthOfStay']=='6 days','LengthOfStay'] = 6
df_claims.loc[df_claims['LengthOfStay']=='1- 2 weeks','LengthOfStay'] = 11
df_claims.loc[df_claims['LengthOfStay']=='2- 4 weeks','LengthOfStay'] = 21
df_claims.loc[df_claims['LengthOfStay']=='4- 8 weeks','LengthOfStay'] = 42
df_claims.loc[df_claims['LengthOfStay']=='26+ weeks','LengthOfStay'] = 180
df_claims['LengthOfStay'].fillna(0, inplace=True)
df_claims['LengthOfStay'] = df_claims['LengthOfStay'].astype(int)

df_claims.dtypes

df_claims['PrimaryConditionGroup_Specialty'] = df_claims['PrimaryConditionGroup'] + '_' + df_claims['Specialty']
df_claims['ProcedureGroup_Specialty'] = df_claims['ProcedureGroup'] + '_' + df_claims['Specialty']
df_claims['ProcedureGroup_PrimaryConditionGroup'] = df_claims['ProcedureGroup'] + '_' + df_claims['PrimaryConditionGroup']
df_claims['PrimaryConditionGroup_PlaceSvc'] = df_claims['PrimaryConditionGroup'] + '_' + df_claims['PlaceSvc']

count_missing(df_claims)

#create a summary table at the member/year level
df_new = pd.DataFrame()
df_group = df_claims.groupby(['Year','MemberID'])
df_new = df_group.agg({
    'ProviderID': ['count','nunique'],
    'Vendor': 'nunique',
    'PCP': 'nunique',
    'PlaceSvc': 'nunique',
    'Specialty': 'nunique',
    'PrimaryConditionGroup': 'nunique',
    'ProcedureGroup': 'nunique',
    'PayDelay': ['sum','max','min', 'mean','std'],
    'LengthOfStay': ['sum','max','min','mean', 'std'],
    'LOS_UNKNOWN': ['sum'],
    'LOS_SUPRESSED': ['sum'],
    'LOS_KNOWN': ['sum'],
    'DSFS': ['sum','max','min','mean', 'std'],
    'CharlsonIndex': ['sum','max','min','mean', 'std'],
    'PrimaryConditionGroup_Specialty': 'nunique',
    'ProcedureGroup_Specialty': 'nunique',
    'ProcedureGroup_PrimaryConditionGroup': 'nunique',
    'PrimaryConditionGroup_PlaceSvc': 'nunique',    
}).reset_index()
df_new.head()

df_group.head()

#Change columns name:
df_new.columns = ['Year', 'MemberID', 'no_Claims', 'no_Providers','no_Vendors','no_PCPs',
                    'no_PlaceSvcs', 'no_Specialities','no_PrimaryConditionGroups','no_ProcedureGroups',
                    'PayDelay_total','PayDelay_max','PayDelay_min','PayDelay_ave','PayDelay_stdev',
                    'LOS_sum','LOS_max','LOS_min','LOS_ave','LOS_std','LOS_TOT_UNKNOWN','LOS_TOT_SUPRESSED','LOS_TOT_KNOWN',
                    'DSFS_sum','DSFS_max','DSFS_min','DSFS_ave','DSFS_std',
                    'CharlsonIndex_sum','CharlsonIndex_max','CharlsonIndex_min','CharlsonIndex_ave','CharlsonIndex_std',
                    'no_PrimaryConditionGroup_Specialty','no_ProcedureGroup_Specialty','no_ProcedureGroup_PrimaryConditionGroup',
                    'no_PrimaryConditionGroup_PlaceSvc'
                 ]
df_new.head()

df_new.info()

count_missing(df_new)

df_new['DSFS_range'] = df_new['DSFS_max'] - df_new['DSFS_min']
df_new['CharlsonIndex_range'] = df_new['CharlsonIndex_max'] - df_new['CharlsonIndex_min']

df_new.loc[(df_new['MemberID']==69258001)]

df_new.shape

"""
for value in df_claims['Specialty'].unique().tolist():  
    df_claims['Specialty_Count_'+value] = (df_claims['Specialty'] == value).astype(int)
    df_agg = df_groupby.agg({'Specialty_Count_'+ value:'sum'}).reset_index()
    df_new = pd.merge(df_new,df_agg, how = 'inner', on = ['MemberID','Year'])

"""

'''
for value in df_claims['PlaceSvc'].unique().tolist():  
    df_claims['PlaceSvc_Count_'+value] = (df_claims['PlaceSvc'] == value).astype(int)
    df_agg = df_groupby.agg({'PlaceSvc_Count_'+ value:'sum'}).reset_index()
    df_new = pd.merge(df_new,df_agg, how = 'inner', on = ['MemberID','Year'])
'''

list_df_col = ['Specialty','PlaceSvc','PrimaryConditionGroup','ProcedureGroup']
for col in list_df_col:
    for value in df_claims[col].unique().tolist():
        df_claims[ str(col) + '_Count_' + str(value)] = (df_claims[col]==value).astype(int)
        df_agg = df_group.agg({str(col) + '_Count_' + str(value):'sum'}).reset_index()
        df_new = pd.merge(df_new, df_agg, how = 'inner', on = ['MemberID','Year'])

df_new.head()

for i in df_new.columns:
    print (i)

df_new.loc[(df_new['MemberID']==18625637)][['Year','MemberID','Specialty_Count_Surgery']]

#df11=df.groupby('key1')['key2'].apply(lambda x: (x=='one').sum()).reset_index(name='count')

'''
df_new['pgc1'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='MSC2a3'))
df_new['pgc2'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='METAB3'))
df_new['pgc3'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='ARTHSPIN'))
df_new['pgc4'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='NEUMENT'))
df_new['pgc5'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='RESPR4'))
df_new['pgc6'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='MISCHRT'))
df_new['pgc7'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='SKNAUT'))
df_new['pgc8'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='GIBLEED'))
df_new['pgc9'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='INFEC4'))
df_new['pgc10'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='TRAUMA'))
'''

'''
df_new['pgc11'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='HEART2'))
df_new['pgc12'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='RENAL3'))
df_new['pgc13'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='ROAMI'))
df_new['pgc14'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='MISCL5'))
df_new['pgc15'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='ODaBNCA'))
df_new['pgc16'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='UTI'))
df_new['pgc17'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='COPD'))
df_new['pgc18'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='GYNEC1'))
df_new['pgc19'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='CANCRB'))
df_new['pgc20'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='FXDISLC'))
'''


'''
df_new['pgc21'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='AMI'))
df_new['pgc22'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='PRGNCY'))
df_new['pgc23'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='HEMTOL'))
df_new['pgc24'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='HEART4'))
df_new['pgc25'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='SEIZURE'))
df_new['pgc26'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='APPCHOL'))
df_new['pgc27'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='CHF'))
df_new['pgc28'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='GYNECA'))
df_new['pgc29'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x.isnull()))
df_new['pgc30'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='PNEUM'))
'''

'''
df_new['pgc31'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='RENAL2'))
df_new['pgc32'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='GIOBSENT'))
df_new['pgc33'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='STROKE'))
df_new['pgc34'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='CANCRA'))
df_new['pgc35'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='FLaELEC'))
df_new['pgc36'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='MISCL1'))
df_new['pgc37'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='HIPFX'))
df_new['pgc38'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='METAB1'))
df_new['pgc39'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='PERVALV'))
df_new['pgc40'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='LIVERDZ'))
'''

'''
df_new['pgc41'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='CATAST'))
df_new['pgc42'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='CANCRM'))
df_new['pgc43'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='PERINTL'))
df_new['pgc44'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='PNCRDZ'))
df_new['pgc45'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='RENAL1'))
df_new['pgc46'] = df_group['PrimaryConditionGroup'].transform(lambda x: sum(x=='SEPSIS'))
'''

'''
df_new['sp1'] = df_group['Specialty'].transform(lambda x: sum(x=='Internal'))
df_new['sp2'] = df_group['Specialty'].transform(lambda x: sum(x=='Laboratory'))
df_new['sp3'] = df_group['Specialty'].transform(lambda x: sum(x=='General Practice'))
df_new['sp4'] = df_group['Specialty'].transform(lambda x: sum(x=='Surgery'))
df_new['sp5'] = df_group['Specialty'].transform(lambda x: sum(x=='Diagnostic Imaging'))
df_new['sp6'] = df_group['Specialty'].transform(lambda x: sum(x=='Emergency'))
df_new['sp7'] = df_group['Specialty'].transform(lambda x: sum(x=='Other'))
df_new['sp8'] = df_group['Specialty'].transform(lambda x: sum(x=='Pediatrics'))
df_new['sp9'] = df_group['Specialty'].transform(lambda x: sum(x=='Rehabilitation'))
df_new['sp10'] = df_group['Specialty'].transform(lambda x: sum(x=='Obstetrics and Gynecology'))
df_new['sp11'] = df_group['Specialty'].transform(lambda x: sum(x=='Anesthesiology'))
df_new['sp12'] = df_group['Specialty'].transform(lambda x: sum(x=='Pathology'))
df_new['sp13'] = df_group['Specialty'].transform(lambda x: sum(x.isnull()))
'''

'''
df_new['pg1'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='EM'))
df_new['pg2'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='PL'))
df_new['pg3'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='MED'))                                                                 
df_new['pg4'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SCS'))
df_new['pg5'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='RAD'))
df_new['pg6'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SDS'))
df_new['pg7'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SIS'))
df_new['pg8'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SMS'))                                                                 
df_new['pg9'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='ANES'))
df_new['pg10'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SGS'))
df_new['pg11'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SEOA'))
df_new['pg12'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SRS'))
df_new['pg13'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SNS'))                                                                 
df_new['pg14'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SAS'))
df_new['pg15'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SUS'))
df_new['pg16'] = df_group['ProcedureGroup'].transform(lambda x: sum(x.isnull()))
df_new['pg17'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SMCD'))
df_new['pg18'] = df_group['ProcedureGroup'].transform(lambda x: sum(x=='SO'))
'''

'''
df_new['ps1'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Office'))                   
df_new['ps2'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Independent Lab'))                                             
df_new['ps3'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Urgent Care'))                                                             
df_new['ps4'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Outpatient Hospital'))                                                                
df_new['ps5'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Inpatient Hospital'))                                                                 
df_new['ps6'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Ambulance'))                                                                 
df_new['ps7'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Other'))                                                                 
df_new['ps8'] = df_group['PlaceSvc'].transform(lambda x: sum(x=='Home'))                                                                 
df_new['ps9'] = df_group['PlaceSvc'].transform(lambda x: sum(x.isnull()))
'''  

df_new.head()

count_missing(df_new)

df_new['PayDelay_stdev'].fillna(-1, inplace=True)
df_new['LOS_std'].fillna(-1, inplace=True)
df_new['DSFS_std'].fillna(-1, inplace=True)
df_new['CharlsonIndex_std'].fillna(-1, inplace=True)

count_missing(df_new)

df_new.to_csv(path_or_buf='New Data/df_Claims_new.csv', index=False)

count_missing(df_new)
