==> 2.Feature Engineering/3_drugs_preprocess.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
import warnings

df_drug = pd.read_csv('../0.data/DrugCount.csv')

df_drug.head()

#Count missing data
def count_missing(df):   
    total = df.isnull().sum().sort_values(ascending=False)
    percent = (df.isnull().sum()/df.isnull().count()).sort_values(ascending=False)
    missing_data  = pd.concat([total, percent], axis=1, keys=['Total', 'Percent'])
    print(missing_data.head(20))
    return missing_data

count_missing(df_drug)

def edit_str(str):
    return int(str.split("-")[0])+1

def edit_count(str):
    str = str.replace("+","")
    return int(str)

df_drug["DSFS"] = df_drug["DSFS"].apply(edit_str)
df_drug["DrugCount"] = df_drug["DrugCount"].apply(edit_count)


df_drug.dtypes

df_group_drug = df_drug.groupby(['Year','MemberID'])

df_new_drug = df_group_drug.agg({'DrugCount': ['sum','max','min','mean','count','std']}).reset_index()

df_new_drug.head()

df_new_drug.columns = ['Year', 'MemberID', 'DrugCount_total', 'DrugCount_max','DrugCount_min','DrugCount_ave','DrugCount_months','DrugCount_std']
df_new_drug.head()

count_missing(df_new_drug)

df_new_drug['DrugCount_std'].fillna(-1, inplace=True)

count_missing(df_new_drug)

df_new_drug.to_csv("new data/df_drugs_new.csv", index=False)
