==> 2.Feature Engineering/5_merge_all.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

df_claims = pd.read_csv('new data/df_claims_new.csv')
df_members = pd.read_csv('new data/df_members_new.csv')
df_labs = pd.read_csv('new data/df_labs_new.csv')
df_drugs = pd.read_csv('new data/df_drugs_new.csv')

df_labs.head()

df_drugs.head()

df_claims.head()

df_members.head()

df_drugs.shape

df_labs.shape

df_lab_drug = pd.merge(df_labs, df_drugs, on=['MemberID', 'Year'], how='outer')
df_lab_drug.head()

df_lab_drug.shape

df_lab_drug[df_lab_drug['MemberID']==210]

#Count missing data
def count_missing(df):   
    total = df.isnull().sum().sort_values(ascending=False)
    percent = (df.isnull().sum()/df.isnull().count()).sort_values(ascending=False)
    missing_data  = pd.concat([total, percent], axis=1, keys=['Total', 'Percent'])
    print(missing_data.head(20))
    return missing_data

count_missing(df_lab_drug)

df_lab_drug_claim = pd.merge(df_lab_drug, df_claims, on=['MemberID', 'Year'], how='outer')
df_lab_drug_claim.head()

missing = count_missing(df_lab_drug_claim)

missing.head(20)

df_lab_drug_claim.shape

df_new_data = pd.merge(df_lab_drug_claim, df_members, on='MemberID')
df_new_data.head()

df_new_data.shape

missing = count_missing(df_new_data)

#df_new_data['LabNull'] = 0
#df_new_data['DrugNull'] = 0
#df_new_data.loc[df_new_data['LabCount_max'].isna(),'LabNull'] = 1
#df_new_data.loc[df_new_data['DrugCount_max'].isna(),'DrugNull'] = 1

df_new_data.fillna(0, inplace=True)

missing = count_missing(df_new_data)

df_new_data.to_csv('new data/new_data_all.csv',index=False)

DaysInHospital_Y2 = pd.read_csv('../0.data/DaysInHospital_Y2.csv')
DaysInHospital_Y2.describe()

DaysInHospital_Y3 = pd.read_csv('../0.data/DaysInHospital_Y3.csv')
DaysInHospital_Y3.describe()

Target = pd.read_csv('../0.data/Target.csv')
Target.describe()

DaysInHospital_Y2['Year'] = 'Y1'
DaysInHospital_Y2['trainset'] = 1
DaysInHospital_Y3['Year'] = 'Y2'
DaysInHospital_Y3['trainset'] = 1
Target['Year'] = 'Y3'
Target['trainset'] = 0

DaysInHospital_Y2.head()

DaysInHospital_Y3.head()

Target.head()

frames = [DaysInHospital_Y2, DaysInHospital_Y3, Target]
DIH = pd.concat(frames)

DIH.rename(columns={'DaysInHospital': 'TARGET'}, inplace=True)

DIH.describe()

DIH.to_csv('new data/DIH_all.csv',index=False)

modelling_set = pd.merge(df_new_data, DIH, on=['MemberID', 'Year'], how='outer')
modelling_set.head()

missing = count_missing(modelling_set)

modelling_set.info

modelling_set[modelling_set['Year'] != 'Y3'].fillna(0, inplace=True)

missing = count_missing(modelling_set)

modelling_set.to_csv('new data/modelling_set.csv',index=False)
